+++ CHANGELOG.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.4.3] - 29.09.2021
+
+## Fixed
+
+* Visibility for `BuildConfig#postcssEnabled` attribute
+* Fix return type of chainable methods with `@returns {this}` in the super class.
+
 ## [1.4.2] - 26.09.2021
 
 ## Fixed
@@ -256,7 +263,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Custom JS module support
 * Basic setup to publish project as NPM package
 
-[Unreleased]: https://github.com/bsi-software/bsi-cx-design-build/compare/1.4.2...HEAD
+[Unreleased]: https://github.com/bsi-software/bsi-cx-design-build/compare/1.4.3...HEAD
+
+[1.4.3]: https://github.com/bsi-software/bsi-cx-design-build/compare/1.4.2...1.4.3
 
 [1.4.2]: https://github.com/bsi-software/bsi-cx-design-build/compare/1.4.1...1.4.2